pyproject.toml
[project]
name = "retailco-analysis"
version = "0.1.0"
description = "Prueba tecnica de analisis de datos y pipeline ETL para RetailCo."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "apache-airflow>=3.0,<4.0",
  "pandas>=2.2,<3.0",
  "psycopg2-binary>=2.9,<3.0",
  "pyarrow>=16.0,<20.0",
  "python-dotenv>=1.0,<2.0",
  "sqlalchemy>=2.0,<3.0",
]

[tool.uv]
package = false
